Make: switch to per CWE (good vs bad) makefiles
The support files are bundled into a lib (juliet-support).
The makefile uses common toolchain variables (e.g. CPP vs CXX). Targets
to make for each specified CWE the individual testcases based on a
Makefile in the CWE's subdirectory. The template Makefile is contained
in this commit. Each testcases good and bad variants are compiled as
distinct executable binary.

Additional target to make clean (WIP) not working if no Makefile with
target clean found in CWEs' subdirectories.